requirements.txt
# Dependencias principales del proyecto
openai>=1.0.0  # Incluye la CLI a partir de la versión 1.x
openai[datalib]  # Para validación de datasets y uso de la CLI
nltk>=3.9.0  # Para métricas BLEU y procesamiento de texto
evaluate>=0.4.0  # Para métricas ROUGE y otras de HuggingFace
rouge_score>=0.1.2  # Necesario para ROUGE en evaluate
absl-py>=2.2.2  # Necesario para ROUGE en evaluate
python-dotenv>=0.19.0
numpy>=1.21.0
requests>=2.26.0
pymupdf>=1.26.0  # Para extracción de texto de PDFs (fitz)
pandas>=2.2.0  # Para manejo y anonimización de CSV
pytest==7.4.3
black==23.11.0
flake8==6.1.0

# Nota: La CLI de OpenAI se usa desde la terminal con la variable de entorno OPENAI_API_KEY
# No es necesario ejecutar 'openai login' en versiones >=1.x 